+++ src/MapWidget.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2020-2022 Tobias Leupold <tl at stonemx dot de>
+// SPDX-FileCopyrightText: 2020-2024 Tobias Leupold <tl at stonemx dot de>
 //
 // SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-KDE-Accepted-GPL
 
@@ -49,6 +49,7 @@ Q_SIGNALS:
     void mapMoved(const Coordinates &center);
     void imagesDropped(const QVector<QString> &paths);
     void requestLoadGpx(const QVector<QString> &paths);
+    void requestAddBookmark();
 
 protected:
     void dragEnterEvent(QDragEnterEvent *event) override;